+++ plugins/amoeba/platforms/reference/src/SimTKReference/AmoebaReferenceHarmonicAngleForce.cpp
@@ -195,7 +195,7 @@ RealOpenMM AmoebaReferenceHarmonicAngleForce::calculateForceAndEnergy( int numAn
                                                                        RealOpenMM angleSextic,
                                                                        vector<RealVec>& forceData) const {
     RealOpenMM energy      = 0.0; 
-    for (unsigned int ii = 0; ii < numAngles; ii++) {
+    for (unsigned int ii = 0; ii < static_cast<unsigned int>(numAngles); ii++) {
         int particle1Index      = particle1[ii];
         int particle2Index      = particle2[ii];
         int particle3Index      = particle3[ii];
